refactor: Move refresh logic into Refresher for easier sharing with AlloyDb
