+++ core/src/main/java/com/google/cloud/sql/core/InstanceData.java
@@ -36,7 +36,7 @@ class InstanceData {
     this.expiration = expiration;
   }
 
-  Instant getExpiration() {
+  public Instant getExpiration() {
     return expiration;
   }
